🐛 FIX: [BE] Character FK 타입 BIGINT 통일

+++ src/models/character.model.js
@@ -3,7 +3,9 @@ const sequelize = require('../config/database');
 
 const Character = sequelize.define('Character', {
   character_id: {
-    type: DataTypes.INTEGER,
+    // stories.character_id와 FK 타입이 정확히 같아야 MySQL이 제약을 생성할 수 있다.
+    // 마이그레이션(0014)도 양쪽 키를 BIGINT로 정의한다.
+    type: DataTypes.BIGINT,
     primaryKey: true,
     autoIncrement: true,
   },
